Revert "Add module version requirements to internal workers modules"
* This reverts commit cce45374874d8a5d30f584dbdc9dad1d0e8c6287.
* Provider passing to child modules is complex and the behavior
changed between Terraform v0.10 and v0.11. We're continuing to
allow both versions so this change should be reverted. For the
time being, those using our internal Terraform modules will have
to be aware of the minimum version for AWS and GCP providers,
there is no good way to do enforcement.